Fixed haversine distance calcs for BT.
Basically vanilla since 1.34 calculates distance using Haversine distances. This means that distances that in-game are further from the equator are scaled down, whereas African ones stay the same. To do this correctly, BT needs to correctly define its equator line in the files, as well as correct latitudes where the map cuts off. The max latitude in BT is the same as vanilla, but the min latitude is different, so that had to change. The equator in BT is also much further north, which meant that BT previously had much shorter distances in the northern parts of the map, while they were much longer in the parts around South Africa.