Nerfed Tax Assessor even more
Clearly I must've done some serious errors in my calcs throwing everything off. New calcs with this current setup results in the tax assessor hitting an equilibrium with an equivalent manufactory in a 10 BT 10 BP 4 ducat TG province, assuming it is built in a province that has gives it double modifier.